+++ examples/typescript/spec/helpers/jasmine-runner.ts
@@ -4,7 +4,7 @@ const Jasmine = require("jasmine");
 import SuiteInfo = jasmine.SuiteInfo;
 
 class CustomProcessor extends DisplayProcessor {
-    public displayJasmineStarted(info: SuiteInfo, log: String): String {
+    public displayJasmineStarted(info: SuiteInfo, log: string): string {
         return `TypeScript ${log}`;
     }
 }