+++ src/display/logger.ts
@@ -2,7 +2,7 @@ import {DisplayProcessor} from "../display-processor";
 import {CustomReporterResult} from "../spec-reporter";
 import SuiteInfo = jasmine.SuiteInfo;
 
-export type ProcessFunction = (displayProcessor: DisplayProcessor, object: ProcessObject, log: String) => String;
+export type ProcessFunction = (displayProcessor: DisplayProcessor, object: ProcessObject, log: string) => string;
 export type ProcessObject = SuiteInfo | CustomReporterResult;
 
 export class Logger {
@@ -10,18 +10,18 @@ export class Logger {
     private currentIndent = "";
     private lastWasNewLine = false;
 
-    constructor(private displayProcessors: DisplayProcessor[], private print: (line: String) => void) {
+    constructor(private displayProcessors: DisplayProcessor[], private print: (line: string) => void) {
     }
 
-    public log(stuff: String): void {
-        stuff.split("\n").forEach((line: String) => {
+    public log(stuff: string): void {
+        stuff.split("\n").forEach((line: string) => {
             this.print(line !== "" ? this.currentIndent + line : line);
         });
         this.lastWasNewLine = false;
     }
 
     public process(object: ProcessObject, processFunction: ProcessFunction): void {
-        let log: String = "";
+        let log = "";
         this.displayProcessors.forEach((displayProcessor: DisplayProcessor) => {
             log = processFunction(displayProcessor, object, log);
         });