+++ src/processors/spec-colors-processor.ts
@@ -2,15 +2,15 @@ import {DisplayProcessor} from "../display-processor";
 import {CustomReporterResult} from "../spec-reporter";
 
 export class SpecColorsProcessor extends DisplayProcessor {
-    public displaySuccessfulSpec(spec: CustomReporterResult, log: String): String {
+    public displaySuccessfulSpec(spec: CustomReporterResult, log: string): string {
         return log.successful;
     }
 
-    public displayFailedSpec(spec: CustomReporterResult, log: String): String {
+    public displayFailedSpec(spec: CustomReporterResult, log: string): string {
         return log.failed;
     }
 
-    public displayPendingSpec(spec: CustomReporterResult, log: String): String {
+    public displayPendingSpec(spec: CustomReporterResult, log: string): string {
         return log.pending;
     }
 }